+++ lib/screens/auth/logout_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:lucide_icons/lucide_icons.dart';
 import 'package:provider/provider.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:terminal_one/widgets/buttons/button3d_primary.dart';
 import 'package:terminal_one/widgets/snackbars/fancy_success_snackbar.dart';
 import 'package:terminal_one/providers/auth_provider.dart';
 import 'package:terminal_one/widgets/buttons/primary_button.dart';
@@ -81,17 +82,19 @@ class LogoutScreen extends StatelessWidget {
             const SizedBox(height: 32),
             
             // Logout Button - Nur so breit wie nötig
-            PrimaryButton(
-              label: 'auth.logout'.tr(),
-              onPressed: () async {
-                // Navigator schließen
-                Navigator.of(context).pop();
-                ScaffoldMessenger.of(context).showSnackBar(
-                  FancySuccessSnackbar.build('auth.logout_success'.tr()),
-                );
-                // Ausloggen
-                await Provider.of<AuthProvider>(context, listen: false).setLoggedIn(false);
-              },
+            IntrinsicWidth(
+              child: PrimaryButton3D(
+                label: 'auth.logout'.tr(),
+                onPressed: () async {
+                  // Navigator schließen
+                  Navigator.of(context).pop();
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    FancySuccessSnackbar.build('auth.logout_success'.tr()),
+                  );
+                  // Ausloggen
+                  await Provider.of<AuthProvider>(context, listen: false).setLoggedIn(false);
+                },
+              ),
             ),
             
             const SizedBox(height: 16),