+++ lib/screens/auth/register_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:terminal_one/widgets/buttons/button3d_primary.dart';
 import '../../utils/responsive_layout.dart';
 import '../../widgets/glassmorphism_scaffold.dart';
 import '../../widgets/app_logo.dart';
@@ -199,10 +200,12 @@ class _RegisterScreenState extends State<RegisterScreen> {
                                       },
                                     label: 'auth.privacy_policy_accept'.tr(),
                                   ),
-                                  PrimaryButton(
-                                    label: 'auth.register_button'.tr(),
-                                    enabled: _canRegister && !_isLoading,
-                                    onPressed: _canRegister && !_isLoading ? _handleRegistration : null,
+                                  IntrinsicWidth(
+                                    child: PrimaryButton3D(
+                                      label: 'auth.register_button'.tr(),
+                                      enabled: _canRegister && !_isLoading,
+                                      onPressed: _canRegister && !_isLoading ? _handleRegistration : null,
+                                    ),
                                   ),
                                 ],
                               ),