+++ lib/screens/games/game_result_screen.dart
@@ -1,5 +1,6 @@
 import 'package:circular_seek_bar/circular_seek_bar.dart';
 import 'package:flutter/material.dart';
+import 'package:terminal_one/widgets/buttons/button3d_primary.dart';
 import 'package:terminal_one/widgets/buttons/primary_button.dart';
 import 'package:terminal_one/screens/home/home_screen_router.dart';
 import '../../widgets/glassmorphism_scaffold.dart';
@@ -118,18 +119,20 @@ class _GameResultScreenState extends State<GameResultScreen> {
                 ),
               ),
               Center(
-                child: PrimaryButton(
-                  label: 'Home',
-                  onPressed: () {
-                      Navigator.pushAndRemoveUntil(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => const HomeScreenRouter(),
-                        ),
-                        (route) => false,
-                      );
-                    },
-                  )
+                child: IntrinsicWidth(
+                  child: PrimaryButton3D(
+                    label: 'Home',
+                    onPressed: () {
+                        Navigator.pushAndRemoveUntil(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => const HomeScreenRouter(),
+                          ),
+                          (route) => false,
+                        );
+                      },
+                    ),
+                )
                 ),
             ],
           ),