+++ lib/widgets/game/button_group.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:lucide_icons/lucide_icons.dart';
+import 'package:terminal_one/widgets/buttons/button3d_primary.dart';
 import 'package:terminal_one/widgets/buttons/ghost_button.dart';
 import 'package:terminal_one/widgets/buttons/primary_button.dart';
 import 'package:terminal_one/widgets/game/button_flippable.dart';
@@ -305,11 +306,13 @@ class _QuizButtonsGroupState extends State<QuizButtonsGroup> {
                           widget.callbackExplanationClicked(question, explanation, hasKI);
                         },
                       ),
-                    PrimaryButton(
-                      leading: LucideIcons.stepForward,
-                      label: isLastQuestion ? 'Zum Ergebnis' : 'Nächste Frage',
-                      onPressed: !isLastQuestion ? _onNextQuestionPressed : _onShowResultPressed,
-                      enabled: true,
+                    IntrinsicWidth(
+                      child: PrimaryButton3D(
+                        leadingIcon: LucideIcons.stepForward,
+                        label: isLastQuestion ? 'Zum Ergebnis' : 'Nächste Frage',
+                        onPressed: !isLastQuestion ? _onNextQuestionPressed : _onShowResultPressed,
+                        enabled: true,
+                      ),
                     ),
                   ],
                 ),